Autocomplete: move utils out of non-stop folder
