+++ vscode/src/non-stop/utils.ts
@@ -1,5 +1,3 @@
-import * as vscode from 'vscode'
-
 export enum CodyTaskState {
     'idle' = 1,
     'working' = 2,
@@ -63,59 +61,4 @@ export const fixupTaskList: CodyTaskList = {
         icon: 'stop',
         description: 'The task failed',
     },
-}
-
-/**
- * Get the last part of the file path after the last slash
- */
-export function getFileNameAfterLastDash(filePath: string): string {
-    const lastDashIndex = filePath.lastIndexOf('/')
-    if (lastDashIndex === -1) {
-        return filePath
-    }
-    return filePath.slice(lastDashIndex + 1)
-}
-
-export function getEditorInsertSpaces(uri: vscode.Uri): boolean {
-    const editor = vscode.window.visibleTextEditors.find(editor => editor.document.uri === uri)
-    if (!editor) {
-        // Default to the same as VS Code default
-        return true
-    }
-
-    const { languageId } = editor.document
-    const languageConfig = vscode.workspace.getConfiguration(`[${languageId}]`, uri)
-    const languageSetting = languageConfig.get('editor.insertSpaces') as boolean | undefined
-    // Prefer language specific setting.
-    const insertSpaces = languageSetting || editor.options.insertSpaces
-
-    // This should never happen: "When getting a text editor's options, this property will always be a boolean (resolved)."
-    if (typeof insertSpaces === 'string' || insertSpaces === undefined) {
-        console.error('Unexpected value when getting "insertSpaces" for the current editor.')
-        return true
-    }
-
-    return insertSpaces
-}
-
-export function getEditorTabSize(uri: vscode.Uri): number {
-    const editor = vscode.window.visibleTextEditors.find(editor => editor.document.uri === uri)
-    if (!editor) {
-        // Default to the same as VS Code default
-        return 4
-    }
-
-    const { languageId } = editor.document
-    const languageConfig = vscode.workspace.getConfiguration(`[${languageId}]`, uri)
-    const languageSetting = languageConfig.get('editor.tabSize') as number | undefined
-    // Prefer language specific setting.
-    const tabSize = languageSetting || editor.options.tabSize
-
-    // This should never happen: "When getting a text editor's options, this property will always be a number (resolved)."
-    if (typeof tabSize === 'string' || tabSize === undefined) {
-        console.error('Unexpected value when getting "tabSize" for the current editor.')
-        return 4
-    }
-
-    return tabSize
 }
